aura-server/requirements.txt
asgiref==3.6.0
Django==3.0
django-dotenv==1.4.2
psycopg2-binary==2.8.6
pytz==2022.7.1
sqlparse==0.4.3
